feat

+++ app/src/main/java/nimonscooked/entity/item/kitchenutensil/KitchenUtensils.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public abstract class KitchenUtensils extends Item {
     protected List<Preparable> contents;
-    
+
     public KitchenUtensils(String id, String name, float x, float y, List<Preparable> contents) {
-        super(id, name, x, y);
+        super(id, name, (int) x, (int) y);
         this.contents = contents;
     }
